Refactor transaction and upgrade/install
This commit refactors the Update method of the transaction
interface. Basically it drops the method from the
interface and the internal steps of the former
Update method are implemented in a new UpdateHelper
interface.

This allows having higher granularity of what call
between Start and Commit transaction in upgrade/install
processes.

In addition, since upgrade/install processes were essentially
diverging on the pratitioning and formatting steps, the
install method is refactored to use the upgrader object
after partitioning.

Transaction interface and upgradehelper interface are
both implemented under the same package as they are both
highly coupled to our snapper based approach.

Signed-off-by: David Cassany <dcassany@suse.com>