Simple 3 way merge for upgrades
This commit implements a simple 3 way merge for the snapshotted
rw volumes (like it happens for /etc). The current implementation
feeds the volume with the contents of the new image and on top of
that it applies any modified file comparing former image content
versus the current content before the upgrade.

The implementation is based on snapper status reports to list modified
and deleted files. Any files listed as a deletion is deleted in target
and any modified file is rsynced to the target. So local modifications
of mutable files will always "win" vs modifications provided by an
upgrade image.

For instance, when you change the contents of /etc in a host, any
added, modified or removed files will be applied to the new snapshot on
upgrades.

Signed-off-by: David Cassany <dcassany@suse.com>