Adding installer disk builder
This commit refactors installer package to be able
to build ISOs and disks. Disks are defined with a very
similar setup to ISOs, they only include the EFI and the
recovery partitions.

Also `build-iso` command is renamed to `build-installer`
as this can be used for both, ISOs and disks. It also
makes prominent the purpose of the built image.

Signed-off-by: David Cassany <dcassany@suse.com>