Add elemental reset command
This commit adds the reset command which installs the system
from the recovery partition to the rest of the current disk.
As of today it can only reset if the rest of the disk is not
partitioned yet.

Running from a installer disk image calling `elemental reset` is
enough to self install and expand to the current disk geomtery. No
further flags are required as all the installation details were
already embedded as part of the build installer process.

Signed-off-by: David Cassany <dcassany@suse.com>